More test updates
