+++ fuel-indexer-tests/Cargo.toml
@@ -24,14 +24,17 @@ chrono = { version = "0.4", features = ["serde"] }
 fuel-indexer = { path = "./../fuel-indexer" }
 fuel-indexer-database = { path = "./../fuel-indexer-database" }
 fuel-indexer-schema = { path = "./../fuel-indexer-schema" }
+fuel-tx = "0.9"
 fuel-types = "0.3"
 fuels = "0.15"
 fuels-abigen-macro = "0.15"
+fuels-core = { version = "0.15" }
 futures = "0.3"
 itertools = "0.10"
 rand = "0.8"
 serde_json = "1.0"
 serde_yaml = "0.8"
+sqlx = "0.6"
 tokio = { version = "1.8", features = ["macros", "rt-multi-thread"] }
 wasmer = "2.0"
 wasmer-compiler-cranelift = { version = "2.0" }
